Add view for `SearchController`
